fix(tests): move tests to single module for CI compatibility
Django test discovery fails with a tests/ subpackage when running
`python manage.py test` without an explicit label. Move to
lupanes/tests.py (standard Django convention).
